ml_operations/requirements.txt
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
catboost==1.2.7
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
MarkupSafe==3.0.2
matplotlib==3.10.0
msgpack==1.1.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
xgboost==2.1.3
